Bug warnings fixes (#466)
* host change

* fixed shadow warning

* made wantedView mobileView a different component

* made wantedView desktopView a component of its own

* added mobile view for wantedSummary

* added Buttons to component file

* created seperate components for bounty tasks

* added codingDesktop

* fixed merge conflicts

* fixed merge conflicts

* fixed merged conflicts and added eslint-disable func-style

* made style props small letters

* restored MODE positon

---------
