+++ frontend/app/src/people/main/header.tsx
@@ -329,9 +329,9 @@ export default function Header() {
                   setShowWelcome(false);
                   goToEditSelf();
                 }}
-                hoverColor={'#5881F8'}
-                activeColor={'#5078F2'}
-                shadowColor={'rgba(97, 138, 255, 0.5)'}
+                hovercolor={'#5881F8'}
+                activecolor={'#5078F2'}
+                shadowcolor={'rgba(97, 138, 255, 0.5)'}
               />
             </Column>
           </div>
