+++ frontend/app/src/people/personViewSlim.tsx
@@ -35,7 +35,6 @@ function makeQR(pubkey: string) {
 
 export default function PersonView(props: any) {
   const { personId, loading, selectPerson, goBack } = props;
-
   // on this screen, there will always be a pubkey in the url, no need for personId
 
   const { main, ui } = useStores();
@@ -64,7 +63,7 @@ export default function PersonView(props: any) {
 
   const people: any = (main.people && main.people.filter((f) => !f.hide)) || [];
 
-  const { id, img, tags, owner_alias, unique_name, price_to_meet, extras, owner_pubkey } =
+  const { id, img, owner_alias, extras, owner_pubkey } =
     person || {};
 
   let { description } = person || {};
@@ -278,15 +277,15 @@ export default function PersonView(props: any) {
 
       const panelStyles = isMobile
         ? {
-            minHeight: 132
-          }
+          minHeight: 132
+        }
         : {
-            maxWidth: 291,
-            minWidth: 291,
-            marginRight: 20,
-            marginBottom: 20,
-            minHeight: 472
-          };
+          maxWidth: 291,
+          minWidth: 291,
+          marginRight: 20,
+          marginBottom: 20,
+          minHeight: 472
+        };
 
       fullSelectedWidget &&
         fullSelectedWidget.forEach((s, i) => {
@@ -298,9 +297,9 @@ export default function PersonView(props: any) {
           const conditionalStyles =
             !isMobile && s?.paid
               ? {
-                  border: '1px solid #dde1e5',
-                  boxShadow: 'none'
-                }
+                border: '1px solid #dde1e5',
+                boxShadow: 'none'
+              }
               : {};
 
           elementArray.push(
@@ -560,13 +559,13 @@ export default function PersonView(props: any) {
                 const hasExtras = extras && extras[name] && extras[name].length > 0;
                 const count: any = hasExtras
                   ? extras[name].filter((f) => {
-                      if ('show' in f) {
-                        // show has a value
-                        if (!f.show) return false;
-                      }
-                      // if no value default to true
-                      return true;
-                    }).length
+                    if ('show' in f) {
+                      // show has a value
+                      if (!f.show) return false;
+                    }
+                    // if no value default to true
+                    return true;
+                  }).length
                   : null;
 
                 return (
@@ -622,8 +621,8 @@ export default function PersonView(props: any) {
   function renderDesktopView() {
     const focusedDesktopModalStyles = newSelectedWidget
       ? {
-          ...tabs[newSelectedWidget]?.modalStyle
-        }
+        ...tabs[newSelectedWidget]?.modalStyle
+      }
       : {};
 
     return (
@@ -840,13 +839,13 @@ export default function PersonView(props: any) {
                 const hasExtras = extras && extras[name] && extras[name].length > 0;
                 const count: any = hasExtras
                   ? extras[name].filter((f) => {
-                      if ('show' in f) {
-                        // show has a value
-                        if (!f.show) return false;
-                      }
-                      // if no value default to true
-                      return true;
-                    }).length
+                    if ('show' in f) {
+                      // show has a value
+                      if (!f.show) return false;
+                    }
+                    // if no value default to true
+                    return true;
+                  }).length
                   : null;
 
                 return (
@@ -958,8 +957,6 @@ export default function PersonView(props: any) {
         visible={showSupport}
         close={() => setShowSupport(false)}
         style={{
-          // top: -64,
-          // height: 'calc(100% + 64px)'
           height: '100%'
         }}
         envStyle={{
@@ -976,10 +973,9 @@ export default function PersonView(props: any) {
                                 subtitle="Because I'm awesome"
                                 buttonlabel="Donate"
                                 defaultinterval="weekly"
-                                imgurl="${
-                                  mediumPic ||
-                                  'https://i.scdn.co/image/28747994a80c78bc2824c2561d101db405926a37'
-                                }"
+                                imgurl="${mediumPic ||
+              'https://i.scdn.co/image/28747994a80c78bc2824c2561d101db405926a37'
+              }"
                             ></sphinx-widget>`
           }}
         />
@@ -1097,25 +1093,6 @@ const Counter = styled.div`
   color: #b0b7bc;
 `;
 
-const ActionInfo = styled.div`
-  font-style: normal;
-  font-weight: normal;
-  font-size: 22px;
-  line-height: 26px;
-  display: flex;
-  align-items: center;
-  text-align: center;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  color: #b0b7bc;
-  margin-bottom: 10px;
-`;
-
-/* Placeholder Text */
-
 const Tabs = styled.div`
   display: flex;
   width: 100%;
@@ -1177,6 +1154,7 @@ const RowWrap = styled.div`
 interface ImageProps {
   readonly src: string;
 }
+
 const Img = styled.div<ImageProps>`
   background-image: url('${(p) => p.src}');
   background-position: center;
