+++ frontend/app/src/people/utils/noneSpaceHomePage.tsx
@@ -31,9 +31,9 @@ export default function NoneSpaceHomePage(props) {
               style={{ marginTop: 20 }}
               onClick={props.action1}
               color="primary"
-              hoverColor={'#5881F8'}
-              activeColor={'#5078F2'}
-              shadowColor={'rgba(97, 138, 255, 0.5)'}
+              hovercolor={'#5881F8'}
+              activecolor={'#5078F2'}
+              shadowcolor={'rgba(97, 138, 255, 0.5)'}
               iconStyle={{
                 top: '13px',
                 right: '14px'
@@ -50,9 +50,9 @@ export default function NoneSpaceHomePage(props) {
               style={{ marginTop: 20, marginLeft: 10 }}
               onClick={props.action2}
               color="success"
-              hoverColor={'#3CBE88'}
-              activeColor={'#2FB379'}
-              shadowColor={'rgba(73, 201, 152, 0.5)'}
+              hovercolor={'#3CBE88'}
+              activecolor={'#2FB379'}
+              shadowcolor={'rgba(73, 201, 152, 0.5)'}
             />
           )}
         </ButtonContainer>
@@ -93,9 +93,9 @@ export default function NoneSpaceHomePage(props) {
             style={{ marginTop: 40 }}
             onClick={props.action1}
             color="primary"
-            hoverColor={'#5881F8'}
-            activeColor={'#5078F2'}
-            shadowColor={'rgba(97, 138, 255, 0.5)'}
+            hovercolor={'#5881F8'}
+            activecolor={'#5078F2'}
+            shadowcolor={'rgba(97, 138, 255, 0.5)'}
           />
         )}
         {props.buttonText2 && (
@@ -107,9 +107,9 @@ export default function NoneSpaceHomePage(props) {
             style={{ marginTop: 20 }}
             onClick={props.action2}
             color="success"
-            hoverColor={'#5881F8'}
-            activeColor={'#5078F2'}
-            shadowColor={'rgba(97, 138, 255, 0.5)'}
+            hovercolor={'#5881F8'}
+            activecolor={'#5078F2'}
+            shadowcolor={'rgba(97, 138, 255, 0.5)'}
           />
         )}
       </ButtonContainerMobile>
