+++ frontend/app/src/people/widgetViews/postBounty/PostBounty.tsx
@@ -7,35 +7,35 @@ import { useStores } from '../../../store';
 import StartUpModal from '../../utils/start_up_modal';
 import { PostModal, PostModalProps } from './PostModal';
 
-interface Props extends Omit<PostModalProps, 'onClose'| 'isOpen'> {
+interface Props extends Omit<PostModalProps, 'onClose' | 'isOpen'> {
   title?: string;
   buttonProps?: {
-    endingIcon?:string;
+    endingIcon?: string;
     leadingIcon?: string;
     color?: 'primary' | 'secondary';
   }
-} 
+}
 
 const color = colors['light'];
 
 const mapBtnColorProps = {
   primary: {
     color: 'success',
-    hoverColor: color.button_primary.hover,
-    activeColor: color.button_primary.active,
-    shadowColor: color.button_primary.shadow,
+    hovercolor: color.button_primary.hover,
+    activecolor: color.button_primary.active,
+    shadowcolor: color.button_primary.shadow,
   },
   secondary: {
     color: 'primary',
-    hoverColor: color.button_secondary.hover,
-    activeColor: color.button_secondary.active,
-    shadowColor: color.button_secondary.shadow,
+    hovercolor: color.button_secondary.hover,
+    activecolor: color.button_secondary.active,
+    shadowcolor: color.button_secondary.shadow,
   }
 }
 
-export const PostBounty: FC<Props> = ( { title= 'Post a Bounty', buttonProps = {
+export const PostBounty: FC<Props> = ({ title = 'Post a Bounty', buttonProps = {
   color: 'primary'
-}, ...modalProps } ) => {
+}, ...modalProps }) => {
 
   const { ui } = useStores();
   const [isOpenPostModal, setIsOpenPostModal] = useState(false);
@@ -43,8 +43,8 @@ export const PostBounty: FC<Props> = ( { title= 'Post a Bounty', buttonProps = {
 
   const isMobile = useIsMobile();
   const showSignIn = () => {
-    if(isMobile) {
-      ui.setShowSignIn(true); 
+    if (isMobile) {
+      ui.setShowSignIn(true);
       return;
     }
     setIsOpenStartupModal(true)
@@ -59,35 +59,35 @@ export const PostBounty: FC<Props> = ( { title= 'Post a Bounty', buttonProps = {
   }
 
   const icon = (() => {
-    if(buttonProps.endingIcon && buttonProps.leadingIcon) {
-      return {leadingIcon: buttonProps.leadingIcon};
+    if (buttonProps.endingIcon && buttonProps.leadingIcon) {
+      return { leadingIcon: buttonProps.leadingIcon };
     }
-    if(buttonProps.leadingIcon) {
-      return {leadingIcon: buttonProps.leadingIcon};
+    if (buttonProps.leadingIcon) {
+      return { leadingIcon: buttonProps.leadingIcon };
     }
-    if(buttonProps.endingIcon) {
-      return {endingIcon: buttonProps.endingIcon};
+    if (buttonProps.endingIcon) {
+      return { endingIcon: buttonProps.endingIcon };
     }
-    return {endingIcon: 'add'};
+    return { endingIcon: 'add' };
   })()
 
 
-  return ( 
-  <>
-    <StyledIconButton
-      {...icon}
-      {...mapBtnColorProps[buttonProps.color || 'primary']}
-      text={title}
-      width={204}
-      height={isMobile ? 36 : 48}
-      iconStyle={iconStyle}
-      onClick={clickHandler}
-    />
-    <PostModal isOpen={isOpenPostModal} onClose={() => setIsOpenPostModal(false)} {...modalProps} /> 
-    {isOpenStartUpModel && (
+  return (
+    <>
+      <StyledIconButton
+        {...icon}
+        {...mapBtnColorProps[buttonProps.color || 'primary']}
+        text={title}
+        width={204}
+        height={isMobile ? 36 : 48}
+        iconStyle={iconStyle}
+        onClick={clickHandler}
+      />
+      <PostModal isOpen={isOpenPostModal} onClose={() => setIsOpenPostModal(false)} {...modalProps} />
+      {isOpenStartUpModel && (
         <StartUpModal closeModal={() => setIsOpenStartupModal(false)} dataObject={'createWork'} buttonColor={'success'} />
       )}
-  </>
+    </>
   )
 }
